backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
anthropic>=0.25.0
APScheduler==3.10.4
